+++ react_on_rails_pro/CHANGELOG.md
@@ -20,6 +20,9 @@ You can find the **package** version numbers from this repo's tags and below in 
     - New `stream_view_containing_react_components` helper method that can be used with `stream_react_component` helper method in react_on_rails gem.
     - Enables progressive page loading and improved performance for server-rendered React components.
   - Added support for replaying console logs that occur during server rendering of streamed React components. This enables debugging of server-side rendering issues by capturing and displaying console output on the client and on the server output. [PR #429](https://github.com/shakacode/react_on_rails_pro/pull/429) by [AbanoubGhadban](https://github.com/AbanoubGhadban).
+  - Added support for handling errors happening during server rendering of streamed React components [PR #432](https://github.com/shakacode/react_on_rails_pro/pull/432) by [AbanoubGhadban](https://github.com/AbanoubGhadban):
+    - It handles errors that happen during the initial render and errors that happen inside suspense boundaries.
+    - Added `raise_non_shell_server_rendering_errors` configuration option.
 - Added support for replaying console logs from asynchronous operations:
   - New `replayServerAsyncOperationLogs` configuration option to enable/disable this feature
   - When enabled, captures and replays console output from async operations during server-side rendering